# Backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SmartDev LMS API Documentation",
        "description": "Comprehensive API documentation for SmartDev Academic Learning Management System. This API provides endpoints for managing courses, students, instructors, assignments, submissions, grades, attendance, and more.",
        "contact": {
            "name": "SmartDev LMS Support",
            "email": "admin@smartdevlms.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://portohansgunawan.my.id",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/v1/courses": {
            "get": {
                "tags": [
                    "Courses"
                ],
                "summary": "Get all courses",
                "description": "Retrieve a list of all courses with instructor information",
                "operationId": "2c728a8c234f4864628b6e0cae94f674",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "course_code": {
                                                "type": "string",
                                                "example": "MTK101"
                                            },
                                            "course_name": {
                                                "type": "string",
                                                "example": "Matematika Dasar"
                                            },
                                            "description": {
                                                "type": "string",
                                                "example": "Mempelajari konsep dasar matematika"
                                            },
                                            "instructor_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "credits": {
                                                "type": "integer",
                                                "example": 3
                                            },
                                            "max_students": {
                                                "type": "integer",
                                                "example": 30
                                            },
                                            "status": {
                                                "type": "string",
                                                "example": "active"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Courses"
                ],
                "summary": "Create new course",
                "description": "Create a new course (Admin/Instructor only)",
                "operationId": "266d569ec244037b25e72e35c4e3b02b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "course_code",
                                    "course_name",
                                    "instructor_id"
                                ],
                                "properties": {
                                    "course_code": {
                                        "type": "string",
                                        "example": "MTK101"
                                    },
                                    "course_name": {
                                        "type": "string",
                                        "example": "Matematika Dasar"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Mempelajari konsep dasar matematika"
                                    },
                                    "instructor_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Course created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "course_code": {
                                            "type": "string",
                                            "example": "MTK101"
                                        },
                                        "course_name": {
                                            "type": "string",
                                            "example": "Matematika Dasar"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/courses/{id}": {
            "get": {
                "tags": [
                    "Courses"
                ],
                "summary": "Get course details",
                "description": "Get detailed information about a specific course including modules, materials, and assignments",
                "operationId": "6374725f9349735b2e90a548e20326b9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Course ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "course_code": {
                                            "type": "string",
                                            "example": "MTK101"
                                        },
                                        "course_name": {
                                            "type": "string",
                                            "example": "Matematika Dasar"
                                        },
                                        "description": {
                                            "type": "string"
                                        },
                                        "instructor": {
                                            "type": "object"
                                        },
                                        "enrollments": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "courseModules": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "assignments": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Course not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Courses"
                ],
                "summary": "Update course",
                "description": "Update an existing course (Admin/Instructor only)",
                "operationId": "0fd7b2aa77580d2dded0f2a448fe8662",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Course ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "course_code": {
                                        "type": "string",
                                        "example": "MTK101"
                                    },
                                    "course_name": {
                                        "type": "string",
                                        "example": "Matematika Dasar"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "instructor_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Course updated successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Course not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your bearer token in the format: Bearer {token}",
                "bearerFormat": "Sanctum Token",
                "scheme": "bearer"
            },
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://portohansgunawan.my.id/oauth/authorize",
                        "tokenUrl": "http://portohansgunawan.my.id/oauth/token",
                        "refreshUrl": "http://portohansgunawan.my.id/token/refresh",
                        "scopes": []
                    }
                }
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}